Set xnack and sram-ecc explicitly per architecture (#145)
Setting these flags explicitly will avoid changing behaviour for
existing architectures when the compiler defaults change.

The architecture-specific flag groups must be protected from CMake's
option de-duplication. The CMake docs for target_compile_options discuss
this problem and the use of SHELL to prevent de-duplication. The
COMPILE_FLAGS property is deprecated, but it seems to be the only way to
avoid de-duplication in CMake 3.11 and earlier.