Workaround for static library size on CI (#270)
Building for all architectures results in a binary that is too large
to link into our test clients. However, the static build is only
tested for a couple architectures. For now, limit the build to those
tested architectures, until we can find a general solution for the
binary size problem.