Remove dependency on cblas (#445)
* Remove dependency on cblas

Use CONFIG mode search by default. FindLAPACK.cmake does not
necessarily set the Fortran runtime libraries required to properly
link LAPACK when it finds netlib-lapack via a module-mode search.
However, it's useful to be able to choose to do a MODULE mode search
as this method supports alternative LAPACK implementations such as
OpenBLAS.

We do use BLAS directly in some cases, but blas is included in the
LAPACK_LIBRARIES anyway. I'm not sure that is guaranteed for all
LAPACK implementations that may be found via module-mode search, but
it can be expected to work that way for the copy of netlib-lapack
that will be found by the config-mode search (which is default).

* Update CHANGELOG.md