Rename cblas_* functions to cpu_* (#482)
* Rename cblas_ prefix to cpu_

The function names for the BLAS / LAPACK wrappers were a bit misleading
as the actual cblas library has not been used in rocSOLVER since
88cae17ddc8bb43ec1ba3f8fde1263cb268a5249.

This change also drops the explicit template type from the calls to
these functions. In nearly all cases, it was superfluous.

* Drop commented-out function implementations

The commented function calls being deleted in this commit were
implemented using the actual cblas library, so they would would
require changes to work after being uncommented. They might as well
just be created when we need them.

* Fix template type deduction

* Address code review comments