Fix debug build for CSRRF tests (#615)
Root cause: Tests for Cholesky-based CSRRF_ANALYSIS and CSRRF_WORKFLOW did not read the actual value of nnzM, causing read_matrix to read too many values from the respective files. The asserts in read_matrix only apply to the debug build, so only the debug build was failing. (Note that nnzM being too large does not cause any issues because it specifies the storage capacity for indA and valA; the actual number of non-zero values is stored in ptrA)

* Fix debug build for CSRRF tests

* Throw exceptions from read_matrix instead of using asserts

* Apply clang format