Add rocsolver team as code owners for docs (#668)
When a file matches multiple patterns in the CODEOWNERS file, it is
assigned the code owners listed in the last matching pattern (not
the union of all code owners in all matching patterns). This means that
the additions to CODEOWNERS in d1cb53f6cdce3e60c98cfb7ad2efe65c3e5b2f78
implicitly removed the rocsolver team members from documentation files.